# GREENALPHA/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GreenAlpha Carbon Footprint Engine",
    "description": "Advanced carbon footprint calculation and sustainability analytics API with <500ms response time",
    "version": "1.0.0"
  },
  "paths": {
...
    "/transport/optimize/route": {
      "post": {
        "tags": [
          "Transport Optimization"
        ],
        "summary": "Optimize Single Route",
        "description": "Get optimal transport routes for a single origin-destination pair",
        "operationId": "optimize_single_route_transport_optimize_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transport/optimize/multimodal": {
      "post": {
        "tags": [
          "Transport Optimization"
        ],
        "summary": "Optimize Multimodal Routes",
        "description": "Find optimal multimodal routes using intermediate hubs",
        "operationId": "optimize_multimodal_routes_transport_optimize_multimodal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transport/optimize/carbon-savings": {
      "post": {
        "tags": [
          "Transport Optimization"
        ],
        "summary": "Calculate Carbon Savings",
        "description": "Calculate potential carbon savings by choosing optimal transport",
        "operationId": "calculate_carbon_savings_transport_optimize_carbon_savings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/transport/optimize/batch": {
      "post": {
        "tags": [
          "Transport Optimization"
        ],
        "summary": "Optimize Batch Routes",
        "description": "Optimize multiple routes in a single request",
        "operationId": "optimize_batch_routes_transport_optimize_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MultiRouteRequest": {
        "properties": {
          "routes": {
            "items": {
              "$ref": "#/components/schemas/RouteRequest"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Routes"
          }
        },
        "type": "object",
        "required": [
          "routes"
        ],
        "title": "MultiRouteRequest",
        "description": "Request model for multiple route comparison"
      },
...
      "RouteRequest": {
        "properties": {
          "origin": {
            "type": "string",
            "title": "Origin",
            "description": "Origin location name"
          },
          "destination": {
            "type": "string",
            "title": "Destination",
            "description": "Destination location name"
          },
          "weight_tonnes": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Tonnes",
            "description": "Shipment weight in tonnes",
            "default": 1.0
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "Optimization priority: carbon, cost, time, balanced",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "RouteRequest",
        "description": "Request model for route optimization"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}